tests: split out setup_lndk fn for reuse
